# schema/api.yaml
openapi: 3.0.3
info:
  title: Arch Blog API
  description: A modern blog platform backend API
  version: 1.0.0
  contact:
    name: API Support
    email: support@archblog.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.archblog.com/v1
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: object
          additionalProperties: true
...
    PaginationMeta:
      type: object
      required:
        - totalItems
        - totalPages
        - currentPage
        - itemsPerPage
      properties:
        totalItems:
          type: integer
          minimum: 0
          example: 100
        totalPages:
          type: integer
          minimum: 0
          example: 10
        currentPage:
          type: integer
          minimum: 1
          example: 1
        itemsPerPage:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
...
    Theme:
      type: object
      required:
        - id
        - name
        - description
        - slug
        - curatorId
        - isActive
        - articleCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Best Practices in Software Architecture"
        description:
          type: string
          maxLength: 500
          example: "A curated collection of posts about software architecture best practices"
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 100
          example: "best-practices-software-architecture"
        curatorId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        isActive:
          type: boolean
          example: true
        articleCount:
          type: integer
          minimum: 0
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    ThemeWithArticles:
      allOf:
        - $ref: '#/components/schemas/Theme'
        - type: object
          required:
            - articles
          properties:
            articles:
              type: array
              items:
                $ref: '#/components/schemas/ThemeArticle'
    ThemeArticle:
      type: object
      required:
        - postId
        - position
        - addedBy
        - addedAt
      properties:
        postId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        position:
          type: integer
          minimum: 0
          example: 0
        addedBy:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        addedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    ThemeSummary:
      type: object
      required:
        - id
        - name
        - description
        - slug
        - curatorId
        - isActive
        - articleCount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Best Practices in Software Architecture"
        description:
          type: string
          example: "A curated collection of posts about software architecture best practices"
        slug:
          type: string
          example: "best-practices-software-architecture"
        curatorId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        isActive:
          type: boolean
          example: true
        articleCount:
          type: integer
          minimum: 0
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    CreateThemeRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Best Practices in Software Architecture"
        description:
          type: string
          maxLength: 500
          example: "A curated collection of posts about software architecture best practices"
    UpdateThemeRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Updated: Best Practices"
        description:
          type: string
          maxLength: 500
          example: "An updated collection of best practices"
    AddArticleRequest:
      type: object
      required:
        - postId
      properties:
        postId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
    ReorderArticlesRequest:
      type: object
      required:
        - postIds
      properties:
        postIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["123e4567-e89b-12d3-a456-426614174000", "987e6543-e21b-12d3-a456-426614174000"]
    PaginatedThemes:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ThemeSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Missing or invalid authentication token"
    ForbiddenError:
      description: User lacks the required permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "You do not have permission to perform this action"
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource not found"
    ValidationError:
      description: The request data is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input data"
            details:
              username: "Username must be between 3 and 30 characters"
    ConflictError:
      description: The request conflicts with existing data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "conflict"
            message: "Username already exists"
    InternalServerError:
      description: An unexpected error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "internal_server_error"
            message: "An unexpected error occurred"
paths:
...
  /themes:
    get:
      tags:
        - Themes
      summary: List themes
      description: Returns a paginated list of themes
      operationId: listThemes
      security: []  # Public endpoint
      parameters:
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: curatorId
          in: query
          description: Filter by curator ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, updated_at, name, article_count]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of themes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedThemes'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Themes
      summary: Create a new theme
      description: Creates a new theme for curating posts
      operationId: createTheme
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThemeRequest'
      responses:
        '201':
          description: Theme created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theme'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /themes/{id}:
    get:
      tags:
        - Themes
      summary: Get a theme by ID
      description: Returns a single theme without articles
      operationId: getTheme
      security: []  # Public endpoint
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Theme retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theme'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Themes
      summary: Update a theme
      description: Updates an existing theme's details
      operationId: updateTheme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateThemeRequest'
      responses:
        '200':
          description: Theme updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theme'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Themes
      summary: Delete a theme
      description: Permanently deletes a theme and all its article associations
      operationId: deleteTheme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Theme deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /themes/slug/{slug}:
    get:
      tags:
        - Themes
      summary: Get a theme by slug
      description: Returns a single theme by its URL slug
      operationId: getThemeBySlug
      security: []  # Public endpoint
      parameters:
        - name: slug
          in: path
          required: true
          description: The URL slug of the theme
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      responses:
        '200':
          description: Theme retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theme'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /themes/{id}/articles:
    get:
      tags:
        - Themes
      summary: Get theme with articles
      description: Returns a theme with all its articles
      operationId: getThemeWithArticles
      security: []  # Public endpoint
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Theme with articles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeWithArticles'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Themes
      summary: Add article to theme
      description: Adds a published post to a theme
      operationId: addArticleToTheme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddArticleRequest'
      responses:
        '201':
          description: Article added successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Themes
      summary: Reorder theme articles
      description: Changes the order of articles in a theme
      operationId: reorderThemeArticles
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderArticlesRequest'
      responses:
        '200':
          description: Articles reordered successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /themes/{id}/articles/{postId}:
    delete:
      tags:
        - Themes
      summary: Remove article from theme
      description: Removes a post from a theme
      operationId: removeArticleFromTheme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme
          schema:
            type: string
            format: uuid
        - name: postId
          in: path
          required: true
          description: The ID of the post to remove
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Article removed successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /themes/{id}/activate:
    post:
      tags:
        - Themes
      summary: Activate a theme
      description: Makes a theme active and visible
      operationId: activateTheme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme to activate
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Theme activated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /themes/{id}/deactivate:
    post:
      tags:
        - Themes
      summary: Deactivate a theme
      description: Makes a theme inactive and hidden
      operationId: deactivateTheme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the theme to deactivate
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Theme deactivated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...